requirements.txt
PyMySQL
flask
flask-peewee
flask-login
flask-bcrypt
flask-wtf
